Создание, удаление курса в admin Просмотр курсов для пользователей

+++ src/main/java/dev/diplom/school/user/repository/UserRepository.java
@@ -1,13 +1,13 @@
 package dev.diplom.school.user.repository;
 
 
-import java.util.Optional;
-
 import dev.diplom.school.user.model.entity.Role;
 import dev.diplom.school.user.model.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserRepository extends JpaRepository<User, Integer> {
+import java.util.Optional;
+
+public interface UserRepository extends JpaRepository<User, Long> {
 
     Optional<User> findByLogin(String login);
